requirements.txt
discord==2.1.0
discord.py==2.1.0
python-dotenv==0.21.0
asyncio==3.4.3